Implement backspace button for cancelling input
Create a new backspace button that will undo the user
last input if they clicked the wrong number.

Fix bug with decimal point where undoing a '.' button click
would not allow the user to add the '.' again in the new number

Fix bug where negation was not properly set when deleting the
negation operator with backspace.

Fix bug where using backspace on a number that has been percented
would give unexpected results to the display.